Add files via upload
The tibyweb server with Perl. This is for Windows, but consists of a web server with Perl that requires no installation. All you do is add your Perl files and have a totally portable web server on an USB thumb drive or SD card. Purely based on the work of others, which is documented in the index.html file.